sManage Comment button and modal separately
